[DEVX-638] Updated the user context.
