+++ setContext.test.js
@@ -1,0 +1,19 @@
+const setContext = require('./setContext');
+
+const req = {
+    cookies: {
+        'express:sess': 'eyJwYXNzcG9ydCI6eyJ1c2VyIjp7InVzZXJpbmZvIjp7InN1YiI6IjAwdTIzOG0yYmVkNUxvaHhXMzU3IiwiZW1haWwiOiJiYWJhamlkZS5iYW5qb0BmdC5jb20iLCJlbWFpbF92ZXJpZmllZCI6dHJ1ZSwibGFzdF9uYW1lIjoiQmFuam8iLCJmaXJzdF9uYW1lIjoiQmFiYWppZGUifSwidG9rZW5zIjp7InRva2VuX3R5cGUiOiJCZWFyZXIiLCJleHBpcmVzX2luIjozMDAsImFjY2Vzc190b2tlbiI6ImV5SnJhV1FpT2lKRFRrUmtYemxGUXpZeE5FUmZhVzExYUVSWVMxVXdWMkphVVd0Tk9XVkRaVWRNYTFaMGNsSldUVTVOSWl3aVlXeG5Jam9pVWxNeU5UWWlmUS5leUoyWlhJaU9qRXNJbXAwYVNJNklrRlVMbEJ5VkhnemVsbFVjbWxaUjFNM2JqRlRYMHhOZERGRlh6QnpRbDlJTUdOdVIxVmthVUZWZDI1WldrRXViMkZ5ZVhZek4yNXFOM2xGYTNsc1JtRXpOVFlpTENKcGMzTWlPaUpvZEhSd2N6b3ZMMlowTFhSbGMzUXViMnQwWVM1amIyMHZiMkYxZEdneUwyRjFjekZyWkRJNWFtYzFURWg0YVVaMU16VTNJaXdpWVhWa0lqb2lZWEJwT2k4dmMzTnZJaXdpYVdGMElqb3hOakl6TnpZeU9EVXhMQ0psZUhBaU9qRTJNak0zTmpNeE5URXNJbU5wWkNJNklqQnZZVFJ0WlRWamVXOTNZMUZtZGxkdU16VTNJaXdpZFdsa0lqb2lNREIxTWpNNGJUSmlaV1ExVEc5b2VGY3pOVGNpTENKelkzQWlPbHNpYjNCbGJtbGtJaXdpYjJabWJHbHVaVjloWTJObGMzTWlMQ0p1WVcxbElpd2laVzFoYVd3aVhTd2ljM1ZpSWpvaVltRmlZV3BwWkdVdVltRnVhbTlBWm5RdVkyOXRJbjAuVGhyUnZ3VDZOcWFyREhCVnU3eVFkOHpmaVVEUjcteFZETUFIcGRxRVZTbjNCeXA5eTRRZEJxeWdLLVI2VEZLLUs0RGU4bmlaSzZuT1R4Q0piUzhxdXY4NmlVZFRyLU5IcDB3VGYzUHNNSnljNHB6QndvN0RZNzFQbHFkNG1NV19RVUJmOUFCcFdYNUs4TXZMdFNMcjFMRTBaOXhIdlZPYV9SdjczTEd1czVGNTJMX25iMHJTSldFMUlidmRqNjdWYkNVVldhT0QwSGNDYnRXZUhvTXhMZnlYQ1NPYnpRbkVLQlFVUHZmeWJIcld0dTZWazBjLXk4QmctS2l3UFVTM3JFNGxnaDBKRTZHaHZZV2RIdWhSWElhcXg4ZDBRMnk5UFNhWGpJUDlnWVZtdVo1WEt4eFRGdXFkVEVJYW9qbnowX1dwV2stUVl6UllOTWc5MFhNTTZnIiwic2NvcGUiOiJvcGVuaWQgb2ZmbGluZV9hY2Nlc3MgbmFtZSBlbWFpbCIsInJlZnJlc2hfdG9rZW4iOiJSYkpFb1U5aXlodlkwRTNsYTk3YlFWSkliYXB5cWd5SnVLOWFpRzBwdlFZIiwiaWRfdG9rZW4iOiJleUpyYVdRaU9pSkRUa1JrWHpsRlF6WXhORVJmYVcxMWFFUllTMVV3VjJKYVVXdE5PV1ZEWlVkTWExWjBjbEpXVFU1Tklpd2lZV3huSWpvaVVsTXlOVFlpZlEuZXlKemRXSWlPaUl3TUhVeU16aHRNbUpsWkRWTWIyaDRWek0xTnlJc0ltVnRZV2xzSWpvaVltRmlZV3BwWkdVdVltRnVhbTlBWm5RdVkyOXRJaXdpZG1WeUlqb3hMQ0pwYzNNaU9pSm9kSFJ3Y3pvdkwyWjBMWFJsYzNRdWIydDBZUzVqYjIwdmIyRjFkR2d5TDJGMWN6RnJaREk1YW1jMVRFaDRhVVoxTXpVM0lpd2lZWFZrSWpvaU1HOWhORzFsTldONWIzZGpVV1oyVjI0ek5UY2lMQ0pwWVhRaU9qRTJNak0zTmpJNE5URXNJbVY0Y0NJNk1UWXlNemMyTmpRMU1Td2lhblJwSWpvaVNVUXVjVmhtU1VwZmJFRlFaR2Q1VkVKVU5IQkVRWFpEVVdoNVNuTXpTVzlHTUU1VlVtMHlkVWcxV0VKM1FTSXNJbUZ0Y2lJNld5SndkMlFpWFN3aWFXUndJam9pTUc5aE1XSm9jM280TUVFNE5VTnhOazh6TlRjaUxDSmhkWFJvWDNScGJXVWlPakUyTWpNM05qSXpPRGNzSW1GMFgyaGhjMmdpT2lKcmVFWnJTM2RIY0hWNVlqVmtlRlJVYjA1TFpWSkJJbjAuVDVCWVRVcFVtQXNETG1aX0tzYnIzV2Y2V0VfbTEwSmp2cFFTQnlVZ2RzRW5lblM1LTlBMGdZbkRmZllJODNBX2JINHBDbzRrWExIN1kzeVVwZTdyTVFURnEyclBsY0JYMzBvZ21vRGdidWxQdmZHdGNhQ2poM3FXY1V3dHNyZXhsX19hMk1PV29LbXlOT2ptSEU5S3l3Z1UzZFB2NUR5Z0RNVEE5ME03UFFBc0R5c2Q4WXJzcTN4cTZUT3JUSExIRU91V2ZyTTVVU0FsR2NVQzVOSGcwT0dmSURmX0xnUG5XVFNHakhNV2p1TXNuYmRMNlo4UzVFY1B5TmpvTEdfVXlndWpCcERrd0k3ekRMTlBVeGJoQlNJUW9PYllKbUJib2xla1NNV3Z3WVBKUThkYm0yX0tzRk9HQzlsblg3WVBFc0ZMSjJTWDBfeC1LYUtWUFl3eFpBIn19fX0='
+    }
+}
+
+const expected = {
+    access_token: 'eyJraWQiOiJDTkRkXzlFQzYxNERfaW11aERYS1UwV2JaUWtNOWVDZUdMa1Z0clJWTU5NIiwiYWxnIjoiUlMyNTYifQ.eyJ2ZXIiOjEsImp0aSI6IkFULlByVHgzellUcmlZR1M3bjFTX0xNdDFFXzBzQl9IMGNuR1VkaUFVd25ZWkEub2FyeXYzN25qN3lFa3lsRmEzNTYiLCJpc3MiOiJodHRwczovL2Z0LXRlc3Qub2t0YS5jb20vb2F1dGgyL2F1czFrZDI5amc1TEh4aUZ1MzU3IiwiYXVkIjoiYXBpOi8vc3NvIiwiaWF0IjoxNjIzNzYyODUxLCJleHAiOjE2MjM3NjMxNTEsImNpZCI6IjBvYTRtZTVjeW93Y1FmdlduMzU3IiwidWlkIjoiMDB1MjM4bTJiZWQ1TG9oeFczNTciLCJzY3AiOlsib3BlbmlkIiwib2ZmbGluZV9hY2Nlc3MiLCJuYW1lIiwiZW1haWwiXSwic3ViIjoiYmFiYWppZGUuYmFuam9AZnQuY29tIn0.ThrRvwT6NqarDHBVu7yQd8zfiUDR7-xVDMAHpdqEVSn3Byp9y4QdBqygK-R6TFK-K4De8niZK6nOTxCJbS8quv86iUdTr-NHp0wTf3PsMJyc4pzBwo7DY71Plqd4mMW_QUBf9ABpWX5K8MvLtSLr1LE0Z9xHvVOa_Rv73LGus5F52L_nb0rSJWE1Ibvdj67VbCUVWaOD0HcCbtWeHoMxLfyXCSObzQnEKBQUPvfybHrWtu6Vk0c-y8Bg-KiwPUS3rE4lgh0JE6GhvYWdHuhRXIaqx8d0Q2y9PSaXjIP9gYVmuZ5XKxxTFuqdTEIaojnz0_WpWk-QYzRYNMg90XMM6g',
+    id_token: 'eyJraWQiOiJDTkRkXzlFQzYxNERfaW11aERYS1UwV2JaUWtNOWVDZUdMa1Z0clJWTU5NIiwiYWxnIjoiUlMyNTYifQ.eyJzdWIiOiIwMHUyMzhtMmJlZDVMb2h4VzM1NyIsImVtYWlsIjoiYmFiYWppZGUuYmFuam9AZnQuY29tIiwidmVyIjoxLCJpc3MiOiJodHRwczovL2Z0LXRlc3Qub2t0YS5jb20vb2F1dGgyL2F1czFrZDI5amc1TEh4aUZ1MzU3IiwiYXVkIjoiMG9hNG1lNWN5b3djUWZ2V24zNTciLCJpYXQiOjE2MjM3NjI4NTEsImV4cCI6MTYyMzc2NjQ1MSwianRpIjoiSUQucVhmSUpfbEFQZGd5VEJUNHBEQXZDUWh5SnMzSW9GME5VUm0ydUg1WEJ3QSIsImFtciI6WyJwd2QiXSwiaWRwIjoiMG9hMWJoc3o4MEE4NUNxNk8zNTciLCJhdXRoX3RpbWUiOjE2MjM3NjIzODcsImF0X2hhc2giOiJreEZrS3dHcHV5YjVkeFRUb05LZVJBIn0.T5BYTUpUmAsDLmZ_Ksbr3Wf6WE_m10JjvpQSByUgdsEnenS5-9A0gYnDffYI83A_bH4pCo4kXLH7Y3yUpe7rMQTFq2rPlcBX30ogmoDgbulPvfGtcaCjh3qWcUwtsrexl__a2MOWoKmyNOjmHE9KywgU3dPv5DygDMTA90M7PQAsDysd8Yrsq3xq6TOrTHLHEOuWfrM5USAlGcUC5NHg0OGfIDf_LgPnWTSGjHMWjuMsnbdL6Z8S5EcPyNjoLG_UygujBpDkwI7zDLNPUxbhBSIQoObYJmBbolekSMWvwYPJQ8dbm2_KsFOGC9lnX7YPEsFLJ2SX0_x-KaKVPYwxZA'
+}
+
+test('sets the user context.', () => {
+    const userTokens = setContext(req);
+
+    expect(userTokens.id_token).toBe(expected.id_token);
+    expect(userTokens.access_token).toBe(expected.access_token);
+});
